Decimal separator (#1)
* Adjust values across the application to the locale used

* Remove package lock file
